Move `Source` related code to separate file
